requirements.txt
django-rest-framework-oauth
pymysql